requirements.txt
torch
torchvision
timm
transformers
albumentations
opencv-python
numpy
pandas
scikit-learn
matplotlib
seaborn
tqdm
Pillow
